Handle sub-zero avg_wra.
Note the definition of avg_wra: the average of ra /(ra_err**2). Thus, we
typically end up with a small number. For example, for a source with RA of 350
deg and typical error 10 arcsec, avg_wra would be 3.5.

When calculating avg_wra for sources near the meridian, we first add a
normalized 180 degrees (that is, 180/(ra_err**2)). We perform the relevant
calculation, then subtract our normalized 180 degrees again.

For sources with 0 <= RA < 180, this works fine. For sources with 180 < RA <
360 this is a problem: when we subtract our normalized 180 degrees, we end up
with a small negative avg_wra.

Usually, this isn't a problem: if we have a negative RA, we take modulus 360,
and we have a positive RA. In this case, though, we can't do modulus 360: we
need modulus 360 * mean(1/ra_err**2).

Unfortunately, the MOD function in SQL does not do what one might hope modulus
would do. Instead, we use the relationship:

  m % n == m - n * floor(m/n)

Closes #4640.

Conflicts:
	tkp/database/utils/associations.py